Update bounds.py align_to_cell_lines() to match FUSION alignment
Modified code to align cell centers with multiple of resolution. It might be better to add an option to the database for pixel-is-point or pixel-is-cell and rename the new version of align_to_cell_lines() to align_to_cell_centers(). There are just a few places where the function is called that would need to add code to use the db option.